+++ logics/classes/propositional/proof_theories/axiom_system.py
@@ -57,7 +57,7 @@ class AxiomSystem:
             inference
         return_error_list: bool
             If False, will just return True or False (exits when it finds an error, more efficient) If True, will return
-            a tuple (boolean, [error_list]) (computes all errors, does not exit on the first, less efficient)
+            (True, a list of ``logics.classes.errors.CorrectionError``)
         exit_on_first_error: bool, optional
             If `return_error_list` and this are both true, it will return a list with a single error instead of many.
             More efficient, since it makes early exits.