+++ logics/classes/propositional/proof_theories/tableaux.py
@@ -564,9 +564,8 @@ class TableauxSystem:
             with no justification are either premises or the negation of the conclusion (this behavior can be overriden
             for other systems, see the source code).
         return_error_list: bool, optional
-            If False, will just return True or False (exits when it finds an error, more efficient). If True, will
-            return a tuple (boolean, [error_list]) (computes all errors, does not exit on the first, less efficient).
-            The error list contains tuples (index path to node, str), see the examples below.
+            If False, will just return True or False (exits when it finds an error, more efficient) If True, will return
+            (True, a list of ``logics.classes.errors.CorrectionError``)
         exit_on_first_error: bool, optional
             If `return_error_list` and this are both true, it will return a list with a single error instead of many.
             More efficient, since it makes early exits.